rename package to othello
