+++ othello_kotlin/app/src/main/kotlin/othello/Othello.kt
@@ -1,4 +1,4 @@
-package othello_kotlin
+package othello
 
 import java.awt.Color
 import kotlin.system.exitProcess
