+++ othello_kotlin/app/src/test/kotlin/othello/OthelloTest.kt
@@ -1,10 +1,9 @@
 /*
  * This Kotlin source file was generated by the Gradle 'init' task.
  */
-package othello_kotlin
+package othello
 
 import kotlin.test.Test
-import kotlin.test.assertNotNull
 
 class OthelloTest {
     @Test fun appHasAGreeting() {
