+++ othello_kotlin/settings.gradle.kts
@@ -13,5 +13,5 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.4.0"
 }
 
-rootProject.name = "othello_kotlin"
+rootProject.name = "othello"
 include("app")
